API/requirements.txt
aiodns==3.1.1
aiofile==3.8.8
aiofiles==23.2.1
aiohttp==3.9.1
aiosignal==1.3.1
async-timeout==4.0.3
asyncpg==0.29.0
attrs==23.2.0
blinker==1.7.0
buildpg==0.4
caio==0.9.13
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptofeed==2.4.0
exceptiongroup==1.2.0
flask==3.0.2
frozenlist==1.4.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
hypercorn==0.16.0
hyperframe==6.0.1
idna==3.6
importlib-metadata==7.0.1
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
multidict==6.0.5
order-book==0.6.0
priority==2.0.0
psycopg2-binary==2.9.9
pycares==4.4.0
pycparser==2.21
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
quart==0.19.4
quart-cors==0.7.0
quart-db==0.8.1
requests==2.31.0
six==1.16.0
sortedcontainers==2.4.0
taskgroup==0.0.0a4
tomli==2.0.1
typing-extensions==4.10.0
urllib3==2.2.1
uvloop==0.19.0
websockets==12.0
werkzeug==3.0.1
wsproto==1.2.0
yapic.json==1.9.1
yarl==1.9.4
zipp==3.17.0
